refactor: move trigger app wiring out of cmd shim
